backend/requirements.txt
fastapi
uvicorn
requests
python-dotenv
fastapi-cors
google-generativeai